refactorize code and do shop endpoints

+++ OrganList/src/main/java/es/organlist/model/dto/ProductListDTO.java
@@ -2,29 +2,22 @@ package es.organlist.model.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.extern.jackson.Jacksonized;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductListDTO {
+    
     @Schema(description = "Identificador del registro que relaciona los productos con las listas")
-    private BigDecimal id;
+    private Integer id;
 
     @Schema(description =  "producto")
-    private BigDecimal product;
+    private Integer product;
 
     @Schema(description =  "lista")
-    private BigDecimal list;
+    private Integer list;
 
     @Schema(description =  "unidades")
     private int units;
